Start the teleport without await
