+++ Explorer/Assets/DCL/Chat/Commands/ReloadSceneChatCommand.cs
@@ -1,11 +1,14 @@
 ﻿using Arch.Core;
 using Cysharp.Threading.Tasks;
 using DCL.Character.CharacterMotion.Components;
+using DCL.CharacterMotion.Components;
 using DCL.RealmNavigation;
 using ECS.SceneLifeCycle;
 using ECS.SceneLifeCycle.Reporting;
 using System;
 using System.Threading;
+using UnityEngine;
+using Utility;
 
 namespace DCL.Chat.Commands
 {
@@ -56,10 +59,7 @@ namespace DCL.Chat.Commands
                 await UniTask.WaitUntil(sceneReadyCondition, cancellationToken: ct);
 
                 if (!isLocalSceneDevelopmentMode && reloadedScene != null)
-                {
-                    WaitForSceneReadiness sceneReadiness = teleportController.TeleportToSceneSpawnPoint(reloadedScene.SceneData.SceneEntityDefinition, ct);
-                    await sceneReadiness.ToUniTask();
-                }
+                    teleportController.StartTeleportToSpawnPoint(reloadedScene.SceneData.SceneEntityDefinition, ct);
 
                 return reloadedScene != null
                     ? "🟢 Current scene has been reloaded"
